edit profile update
Design update + showmessage and pagereload added
